+++ src/main/java/org/example/controllers/ConfigListController.java
@@ -17,7 +17,16 @@ import org.example.pdf_creator.factories.TextSectionListType;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class ConfigListController {
+/**
+ * Controller for config-list.fxml
+ * Handles: applying listeners on Font Preset List View,
+ * listeners on Text Section List Type choice box,
+ * button for creating a new FontPreset,
+ * and transition to Parent / Child TextSection.
+ * Transition to parent TextSection and choice of TextSectionListType
+ * are hidden for TextSectionList of type PdfCreationConfiguration
+ */
+public class ConfigListController extends MainPanelController {
 
     @FXML
     private Pane listHandlerChoicePane;
@@ -36,42 +45,46 @@ public class ConfigListController {
 
     public void init(TextSectionList currentTextSectionList) {
         this.textSectionList = currentTextSectionList;
-        listHandlerChoiceBox.getItems().addAll(TextSectionListType.GRID, TextSectionListType.LIST);
-        listHandlerChoiceBox.setValue(textSectionList.getTextSectionListType());
-        listHandlerChoiceBox.getSelectionModel().selectedIndexProperty().addListener((ov, value, new_value) -> {
-            this.textSectionList = listHandlerChoiceBox.getItems().get(new_value.intValue()).createList(currentTextSectionList);
-        });
-        ObservableList<String> data = FXCollections.observableArrayList(currentTextSectionList.fontPresets.stream().map(FontPreset::toListViewString).collect(Collectors.toList()));
-        fontPresetListView.setItems(data);
-        fontPresetListView.setOnMouseClicked(click -> {
-            if (click.getClickCount() == 2) {
-                List<FontPreset> fontPresetList = currentTextSectionList.fontPresets;
-                FontPreset preset = fontPresetList.stream().filter(el -> el.equalsString(fontPresetListView.getSelectionModel()
-                        .getSelectedItem())).findFirst().get();
-                FontPresetPickerDialogBox dialogBox = new FontPresetPickerDialogBox(preset.clone(), App.stage, (value) -> {
-                    int index =  fontPresetList.indexOf(fontPresetList.stream().filter(el -> el.getUuid().equals(((FontPreset)value).getUuid())).findFirst().get());
-                    fontPresetList.remove(index);
-                    fontPresetList.add(index, (FontPreset)value);
-                    currentTextSectionList.setFontPresets(fontPresetList);
-                    ObservableList<String> data1 = FXCollections.observableArrayList(fontPresetList.stream().map(FontPreset::toListViewString).collect(Collectors.toList()));
-                    fontPresetListView.setItems(data1);
-                });
-                dialogBox.open();
-            }
-        });
-        newPresetButton.setOnMouseClicked((event) -> {
-            new FontPresetPickerDialogBox(new FontPreset(), App.stage, (value -> {
-                List<FontPreset> fontPresets = this.textSectionList.getFontPresets();
-                fontPresets.add((FontPreset) value);
-                this.textSectionList.setFontPresets(fontPresets);
-                ObservableList<String> data1 = FXCollections.observableArrayList(fontPresets.stream().map(FontPreset::toListViewString).collect(Collectors.toList()));
-                fontPresetListView.setItems(data1);
-
-            })).open();
-        });
-        listHandlerChoicePane.setVisible(!(currentTextSectionList instanceof PdfCreationConfiguration));
+        setUpListHandlerChoiceBox();
+        setUpFontPresetListView();
+        applyListenerOnNewPresetButton();
+        conditionallyCoverUIElementsBasedOnTextSectionListType();
 
     }
 
+    private void setUpListHandlerChoiceBox() {
+        listHandlerChoiceBox.getItems().addAll(TextSectionListType.GRID, TextSectionListType.LIST);
+        listHandlerChoiceBox.setValue(textSectionList.getTextSectionListType());
+        listHandlerChoiceBox.getSelectionModel().selectedIndexProperty().addListener((ov, value, new_value) -> {
+            this.textSectionList = listHandlerChoiceBox.getItems().get(new_value.intValue()).createList(textSectionList);
+        });
+    }
+
+
+    /**
+     * Method for setting up Font Preset List View.
+     * First it grabs all font presets for this text section and puts their string representation into list
+     * then applies a listener on a double click, which opens a FontPresetPickerDialogBox in which
+     * user can edit already existing font preset
+     */
+    private void setUpFontPresetListView() {
+        setUpFontPresetListViewWithParameters(this.fontPresetListView, this.textSectionList);
+    }
+
+    private void applyListenerOnNewPresetButton() {
+        newPresetButton.setOnMouseClicked((event) -> {
+            new FontPresetPickerDialogBox(new FontPreset(), App.stage, (value -> {
+                List<FontPreset> fontPresets = this.textSectionList.getFontPresetList();
+                fontPresets.add((FontPreset) value);
+                this.textSectionList.setFontPresetList(fontPresets);
+                ObservableList<String> data1 = FXCollections.observableArrayList(fontPresets.stream().map(FontPreset::toListViewString).collect(Collectors.toList()));
+                fontPresetListView.setItems(data1);
+            })).open();
+        });
+    }
+
+    private void conditionallyCoverUIElementsBasedOnTextSectionListType() {
+        listHandlerChoicePane.setVisible(!(this.textSectionList instanceof PdfCreationConfiguration));
+    }
 
 }