+++ src/main/java/org/example/pdf_creator/content/FontPresetListContainer.java
@@ -1,0 +1,8 @@
+package org.example.pdf_creator.content;
+
+import java.util.List;
+
+public interface FontPresetListContainer {
+    List<FontPreset> getFontPresetList();
+    void setFontPresetList(List<FontPreset> fontPresetList);
+}