+++ src/main/java/org/example/pdf_creator/content/Subsection.java
@@ -41,15 +41,15 @@ public class Subsection extends TextSection {
         return null;
     }
 
-    public void setFontPresets(List<FontPreset> presets) {
-        this.fontPresets = presets;
+    public void setFontPresetList(List<FontPreset> presets) {
+        this.fontPresetList = presets;
         if(textSectionList != null) {
             textSectionList.conditionallySetFontPresets(getChildrenFontPresets());
         }
     }
     public void setContentList(TextSectionList list) {
         this.textSectionList = list;
-        if(this.fontPresets != null) {
+        if(this.fontPresetList != null) {
             textSectionList.conditionallySetFontPresets(getChildrenFontPresets());
         }
     }