+++ src/main/java/org/example/pdf_creator/content/sectionorganizers/ListedTextSections.java
@@ -69,4 +69,5 @@ public class ListedTextSections extends TextSectionList {
     public TextSectionListType getTextSectionListType() {
         return TextSectionListType.LIST;
     }
+
 }